+++ src/Plugins/JJMasterData.Python/Python/JJMasterData/Core/FormEvents/Args.py
@@ -1,0 +1,18 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class FormAfterActionEventArgs:
+    Values: dict
+    UrlRedirect: str
+
+
+@dataclass
+class FormBeforeActionEventArgs:
+    Values: dict
+    Errors: dict
+
+
+@dataclass
+class MetadataLoadEventArgs:
+    Metadata: any
