--- src/Plugins/JJMasterData.Python/Python/JJMasterData/FormEvents/Args/FormAfterActionEventArgs.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class FormAfterActionEventArgs:
-    Values: dict
-    UrlRedirect: str
