--- src/Plugins/JJMasterData.Python/Python/JJMasterData/FormEvents/Args/FormBeforeActionEventArgs.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class FormBeforeActionEventArgs:
-    Values: dict
-    Errors: dict
