--- src/Plugins/JJMasterData.Python/Python/JJMasterData/FormEvents/Args/MetadataLoadEventArgs.py
@@ -1,6 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class MetadataLoadEventArgs:
-    Metadata: any
